Split the peer selection tests up
The module was rather on the huge side.
